Finished Pong Minigame
Finished the Pong minigame. This minigame is automatically chosen when all players are ready and there are exactly 2 players in the lobby. This is because I did not have enough time to be able to figure out support for more players in the Pong game.
The position syncing has also been changed a bit. It is now being handled directly by the NetworkComponent instead of the SyncManager, this does mean more sporadic and unpredictable packet sending but is ultimately simpler as it is all contained in one script instead of over multiple.
The host is the one responsible for spawning in all the player's duck so that we can be sure they all have the same ObjectID and OwnerID across all the clients. This involved adding a special check to handle spawning in the ducks over and from the network only when they are not already in the scene, but if the object being spawned is not a duck, then it doesn't matter as this special logic is skipped.
We discovered that the pain in our ass that made everythign not work was that our PositionPackets weren't being properly Deserialized. They were calculating ther size of the packet correctly but never updating the bufferSize or offset so our games would simply freeze the second they entered the client or server's buffer.